Update Zeek plugin for JIT changes.
We can now merge the two plugin classes as we no longer need to keep the
JIT functionality separate.
