+++ zeek/plugin/src/config.h.in
@@ -4,8 +4,6 @@
 
 namespace spicy::zeek::configuration {
 
-#cmakedefine ZEEK_HAVE_JIT
-
 // We make sure this is always defined, either as 0 or 1, so that we
 // can catch when this header wasn't included.
 #cmakedefine01 ZEEK_DEBUG_BUILD
